Added/modified test-cases

+++ controllers/reconciler.go
@@ -216,35 +216,39 @@ func (r *SingleClusterReconciler) Reconcile() (ctrl.Result, error) {
 		return reconcile.Result{}, err
 	}
 
-	podList, gErr := r.getClusterPodList()
-	if gErr != nil {
-		r.Log.Error(gErr, "Failed to get cluster pod list")
-		return reconcile.Result{}, gErr
-	}
-
-	r.Log.Info("Try to recover failed/pending pods if any")
-
-	var anyPodFailed bool
-	// Try to recover failed/pending pods by deleting them
-	for idx := range podList.Items {
-		if cErr := utils.CheckPodFailed(&podList.Items[idx]); cErr != nil {
-			anyPodFailed = true
-
-			if err := r.createOrUpdatePodServiceIfNeeded([]string{podList.Items[idx].Name}); err != nil {
-				return reconcile.Result{}, err
-			}
-
-			if err := r.Client.Delete(context.TODO(), &podList.Items[idx]); err != nil {
-				r.Log.Error(err, "Failed to delete pod", "pod", podList.Items[idx].Name)
-				return reconcile.Result{}, err
-			}
-
-			r.Log.Info("Deleted pod", "pod", podList.Items[idx].Name)
+	// Try to recover pods only when MaxIgnorablePods is set
+	if r.aeroCluster.Spec.RackConfig.MaxIgnorablePods != nil {
+		podList, gErr := r.getClusterPodList()
+		if gErr != nil {
+			r.Log.Error(gErr, "Failed to get cluster pod list")
+			return reconcile.Result{}, gErr
 		}
-	}
 
-	if anyPodFailed {
-		return reconcile.Result{Requeue: true}, nil
+		r.Log.Info("Try to recover failed/pending pods if any")
+
+		var anyPodFailed bool
+		// Try to recover failed/pending pods by deleting them
+		for idx := range podList.Items {
+			if cErr := utils.CheckPodFailed(&podList.Items[idx]); cErr != nil {
+				anyPodFailed = true
+
+				if err := r.createOrUpdatePodServiceIfNeeded([]string{podList.Items[idx].Name}); err != nil {
+					return reconcile.Result{}, err
+				}
+
+				if err := r.Client.Delete(context.TODO(), &podList.Items[idx]); err != nil {
+					r.Log.Error(err, "Failed to delete pod", "pod", podList.Items[idx].Name)
+					return reconcile.Result{}, err
+				}
+
+				r.Log.Info("Deleted pod", "pod", podList.Items[idx].Name)
+			}
+		}
+
+		if anyPodFailed {
+			r.Log.Info("Found failed/pending pod(s), requeuing")
+			return reconcile.Result{Requeue: true}, nil
+		}
 	}
 
 	r.Log.Info("Reconcile completed successfully")
